Switch the Windows installer from NSIS to MSI
This means we get a few bugs fixed:

 * We appear in the Add/Remove Programs UI
 * We have an uninstaller that works on Windows 10
 * The installation is pretty, and scriptable
 * We actually install the translations

+++ contrib/ci/build_windows.sh
@@ -1,10 +1,18 @@
 #!/bin/sh
 set -e
+
+# if invoked outside of CI
+if [ "$CI" != "true" ]; then
+	echo "Not running in CI, please manually configure Windows build"
+	exit 1
+fi
+
 #prep
 export LC_ALL=C.UTF-8
 root=$(pwd)
 export DESTDIR=${root}/dist
 build=$root/build-win32
+
 rm -rf $DESTDIR $build
 
 # For logitech bulk controller being disabled (-Dplugin_logitech_bulkcontroller=disabled):
@@ -27,6 +35,9 @@ meson .. \
     -Ddocs=none \
     -Dhsi=false \
     -Dman=false \
+    -Dfish_completion=false \
+    -Dbash_completion=false \
+    -Dfirmware-packager=false \
     -Dmetainfo=false \
     -Dcompat_cli=false \
     -Dsoup_session_compat=false \
@@ -39,14 +50,22 @@ meson .. \
     -Dlibxmlb:gtkdoc=false \
     -Dlibjcat:man=false \
     -Dlibjcat:gpg=false \
+    -Dlibjcat:tests=false \
     -Dlibjcat:introspection=false \
     -Dgusb:tests=false \
     -Dgusb:docs=false \
     -Dgusb:introspection=false \
     -Dgusb:vapi=false $@
 VERSION=$(meson introspect . --projectinfo | jq -r .version)
-ninja -v
-ninja -v install
+
+# run tests
+export WINEPATH="/usr/x86_64-w64-mingw32/sys-root/mingw/bin/;$build/libfwupd/;$build/libfwupdplugin/;$build/subprojects/libxmlb/src/;$build/subprojects/gcab/libgcab/;$build/subprojects/libjcat/libjcat/;$build/subprojects/gusb/gusb/"
+ninja -C $build -v
+ninja -C $build test
+
+# switch to release optimisations
+meson configure -Dtests=false -Dbuildtype=release
+ninja -C $build -v install
 
 #generate news release
 cd $root
@@ -57,25 +76,91 @@ echo $VERSION > $DESTDIR/VERSION
 sed -i 's,UpdateMotd=.*,UpdateMotd=false,' $DESTDIR/etc/fwupd/daemon.conf
 
 # create a setup binary
-cd $DESTDIR
-mkdir -p $DESTDIR/setup
-makensis -NOCD $build/contrib/setup-win32.nsi
+CERTDIR=/etc/pki/tls/certs
+MINGW32BINDIR=/usr/x86_64-w64-mingw32/sys-root/mingw/bin
 
-#so that it's actually executable
-cp /usr/x86_64-w64-mingw32/sys-root/mingw/bin/*.dll bin
+# deps
+find $MINGW32BINDIR \
+	-name curl.exe \
+	-o -name gspawn-win64-helper-console.exe \
+	-o -name gspawn-win64-helper.exe \
+	-o -name iconv.dll \
+	-o -name libarchive-13.dll \
+	-o -name libbrotlicommon.dll \
+	-o -name libbrotlidec.dll \
+	-o -name libbz2-1.dll \
+	-o -name libcrypto-1_1-x64.dll \
+	-o -name libcurl-4.dll \
+	-o -name libffi-*.dll \
+	-o -name libgcc_s_seh-1.dll \
+	-o -name libgio-2.0-0.dll \
+	-o -name libglib-2.0-0.dll \
+	-o -name libgmodule-2.0-0.dll \
+	-o -name libgmp-10.dll \
+	-o -name libgnutls-30.dll \
+	-o -name libgnutls-30.dll \
+	-o -name libgobject-2.0-0.dll \
+	-o -name libhogweed-*.dll \
+	-o -name libidn2-0.dll \
+	-o -name libintl-8.dll \
+	-o -name libjson-glib-1.0-0.dll \
+	-o -name liblzma-5.dll \
+	-o -name libnettle-*.dll \
+	-o -name libp11-kit-0.dll \
+	-o -name libpcre-1.dll \
+	-o -name libsqlite3-0.dll \
+	-o -name libssh2-1.dll \
+	-o -name libssl-1_1-x64.dll \
+	-o -name libssp-0.dll \
+	-o -name libtasn1-6.dll \
+	-o -name libusb-1.0.dll \
+	-o -name libwinpthread-1.dll \
+	-o -name libxml2-2.dll \
+	-o -name zlib1.dll \
+	| wixl-heat \
+	-p $MINGW32BINDIR/ \
+	--directory-ref BINDIR \
+	--var "var.MINGW32BINDIR" \
+	--component-group "CG.fwupd-deps" | \
+	tee $build/contrib/fwupd-deps.wxs
 
-#remove static archives
-find -type f -print0 -name "*.dll.a" | xargs rm -f
+# CA bundle
+echo $CERTDIR/ca-bundle.crt | wixl-heat \
+	-p $CERTDIR/ \
+	--directory-ref BINDIR \
+	--var "var.CRTDIR" \
+	--component-group "CG.fwupd-certs" | \
+	tee $build/contrib/fwupd-certs.wxs
 
-#remove stuff that we really don't need
-rm -fr gcab.exe \
-       xb-tool.exe \
-       share/man \
-       include \
-       fwupd \
-       lib/*.a \
-       lib/pkgconfig/ \
-       var
+# our files
+find $DESTDIR | \
+	wixl-heat \
+	-p $DESTDIR/ \
+	-x include/ \
+	-x share/fwupd/device-tests/ \
+	-x share/tests/ \
+	-x share/man/ \
+	-x share/doc/ \
+	-x lib/pkgconfig/ \
+	--directory-ref INSTALLDIR \
+	--var "var.DESTDIR" \
+	--component-group "CG.fwupd-files" | \
+	tee $build/contrib/fwupd-files.wxs
 
-export WINEPATH="/usr/x86_64-w64-mingw32/sys-root/mingw/bin/;$build/libfwupd/;$build/libfwupdplugin/;$build/subprojects/libxmlb/src/;$build/subprojects/gcab/libgcab/;$build/subprojects/libjcat/libjcat/;$build/subprojects/gusb/gusb/"
-ninja -C $build test
+MSI_FILENAME="$DESTDIR/setup/fwupd-$VERSION-setup-x86_64.msi"
+mkdir -p setup
+wixl -v \
+	$build/contrib/fwupd.wxs \
+	$build/contrib/fwupd-deps.wxs \
+	$build/contrib/fwupd-certs.wxs \
+	$build/contrib/fwupd-files.wxs \
+	-D CRTDIR=$CERTDIR \
+	-D MINGW32BINDIR=$MINGW32BINDIR \
+	-D DESTDIR=$DESTDIR \
+	-o ${MSI_FILENAME}
+
+# check the msi archive can be installed and removed (use "wine uninstaller" to do manually)
+wine msiexec /i ${MSI_FILENAME}
+tree ~/.wine/drive_c/Program\ Files/fwupd/
+wine ~/.wine/drive_c/Program\ Files/fwupd/bin/fwupdtool get-plugins
+wine msiexec /x ${MSI_FILENAME}
